build: added a github action to deploy to server
Will be run every time a push is made to the repo. We can change it to
manual if we start to run out of compute time.

TODO: Im not sure how well the caching works on pnpm install so if we
want deploys to be faster it may be good to review that build step.

Not pictured is the aprox 15 commits I squashed together to get it to
work lmao.