fix: remove calendar title with url modification
I did this because it dosent really matter in the calendar and we don't
want people to send mail there if we decide to use an alternate email in
the contact section. Theres probably a way to hide the team email from
the url but I don't think its an issue.